+++ packages/sdks/src/constants/extra-components.ts
@@ -16,10 +16,10 @@ import { componentInfo as formSelectComponentInfo } from '../blocks/form/select/
 import { default as FormSelect } from '../blocks/form/select/index.js';
 import { componentInfo as formSubmitButtonComponentInfo } from '../blocks/form/submit-button/component-info.js';
 import { default as FormSubmitButton } from '../blocks/form/submit-button/index.js';
+import { componentInfo as textAreaComponentInfo } from '../blocks/form/textarea/component-info.js';
+import { default as Textarea } from '../blocks/form/textarea/index.js';
 import { componentInfo as imgComponentInfo } from '../blocks/img/component-info.js';
 import { default as Img } from '../blocks/img/index.js';
-import { componentInfo as textAreaComponentInfo } from '../blocks/textarea/component-info.js';
-import { default as Textarea } from '../blocks/textarea/index.js';
 import { componentInfo as videoComponentInfo } from '../blocks/video/component-info.js';
 import { default as Video } from '../blocks/video/index.js';
 import type { RegisteredComponent } from '../context/types.js';
